pyproject.toml
[project]
name = "mcp-server-youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "fastmcp>=2.2.5",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
